+++ src/main/java/ch/epfl/leb/sass/ijplugin/GUI.java
@@ -1,5 +1,5 @@
 /* 
- * Copyright (C) 2017 Laboratory of Experimental Biophysics
+ * Copyright (C) 2017-2018 Laboratory of Experimental Biophysics
  * Ecole Polytechnique Federale de Lausanne
  * 
  * Author: Marcel Stefko
@@ -24,6 +24,7 @@ import ij.plugin.frame.PlugInFrame;
 import java.io.File;
 import javax.swing.JFileChooser;
 import javax.swing.filechooser.FileNameExtensionFilter;
+
 import java.io.IOException;
 
 /**