+++ src/main/java/ch/epfl/leb/sass/models/fluorophores/internal/commands/FluorophoreCommandBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Laboratory of Experimental Biophysics
+ * Copyright (C) 2017-2018 Laboratory of Experimental Biophysics
  * Ecole Polytechnique Federale de Lausanne
  * 
  * This program is free software: you can redistribute it and/or modify
@@ -17,6 +17,7 @@
  */
 package ch.epfl.leb.sass.models.fluorophores.internal.commands;
 
+import ch.epfl.leb.sass.models.components.Laser;
 import ch.epfl.leb.sass.models.components.Camera;
 import ch.epfl.leb.sass.models.photophysics.FluorophoreDynamics;
 import ch.epfl.leb.sass.models.psfs.PSFBuilder;
@@ -28,11 +29,44 @@ import ch.epfl.leb.sass.models.psfs.PSFBuilder;
  */
 public interface FluorophoreCommandBuilder {
     
+    /**
+     * Creates a new Command instance for generating a fluorophore distribution.
+     * 
+     * @return The new instance of a FluorophoreCommand.
+     */    
     public FluorophoreCommand build();
     
+    /**
+     * Sets the builder's Camera instance.
+     * 
+     * @param camera The camera used for building Fluorophore distributions.
+     * @return A new copy of the builder.
+     */
     public FluorophoreCommandBuilder camera(Camera camera);
+    
+    /**
+     * Sets the FluorescenceDynamics.
+     * 
+     * @param fluorDynamics The fluorescence dynamical system.
+     * @return A new copy of the builder.
+     */
     public FluorophoreCommandBuilder fluorDynamics(
             FluorophoreDynamics fluorDynamics);
+    
+    /**
+     * Sets the laser that is used to build the fluorophore distributions.
+     * 
+     * @param laser The laser whose properties are used to build fluorophores.
+     * @return A new copy of the builder.
+     */
+    public FluorophoreCommandBuilder laser(Laser laser);
+    
+    /**
+     * Sets the PSF builder that will create the fluorophores' PSFs.
+     * 
+     * @param psfBuilder The PSF builder.
+     * @return A new copy of the builder.
+     */
     public FluorophoreCommandBuilder psfBuilder(PSFBuilder psfBuilder);
     
 }