+++ src/main/java/ch/epfl/leb/sass/models/fluorophores/internal/commands/GenerateFluorophoresFromCSV.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Laboratory of Experimental Biophysics
+ * Copyright (C) 2017-2018 Laboratory of Experimental Biophysics
  * Ecole Polytechnique Federale de Lausanne
  * 
  * This program is free software: you can redistribute it and/or modify
@@ -17,10 +17,12 @@
  */
 package ch.epfl.leb.sass.models.fluorophores.internal.commands;
 
+import ch.epfl.leb.sass.models.psfs.PSFBuilder;
+import ch.epfl.leb.sass.models.components.Laser;
+import ch.epfl.leb.sass.models.components.Camera;
 import ch.epfl.leb.sass.models.fluorophores.Fluorophore;
 import ch.epfl.leb.sass.models.photophysics.FluorophoreDynamics;
-import ch.epfl.leb.sass.models.components.Camera;
-import ch.epfl.leb.sass.models.psfs.PSFBuilder;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -51,6 +53,11 @@ public final class GenerateFluorophoresFromCSV implements FluorophoreCommand {
     private final FluorophoreDynamics fluorDynamics;
     
     /**
+     * The laser that illuminates the fluorophores.
+     */
+    private final Laser laser;
+    
+    /**
      * A builder for creating PSFs.
      */
     private final PSFBuilder psfBuilder;
@@ -67,6 +74,7 @@ public final class GenerateFluorophoresFromCSV implements FluorophoreCommand {
         private File file;
         private Camera camera;
         private FluorophoreDynamics fluorDynamics;
+        private Laser laser;
         private PSFBuilder psfBuilder;
         private boolean rescale;
         
@@ -82,6 +90,11 @@ public final class GenerateFluorophoresFromCSV implements FluorophoreCommand {
         @Override
         public Builder fluorDynamics(FluorophoreDynamics fluorDynamics) {
             this.fluorDynamics = fluorDynamics;
+            return this;
+        }
+        @Override
+        public Builder laser(Laser laser) {
+            this.laser = laser;
             return this;
         }
         @Override
@@ -106,6 +119,7 @@ public final class GenerateFluorophoresFromCSV implements FluorophoreCommand {
      */
     private GenerateFluorophoresFromCSV(Builder builder) {
         this.camera = builder.camera;
+        this.laser = builder.laser;
         this.fluorDynamics = builder.fluorDynamics;
         this.file = builder.file;
         this.psfBuilder = builder.psfBuilder;
@@ -123,6 +137,7 @@ public final class GenerateFluorophoresFromCSV implements FluorophoreCommand {
             return FluorophoreReceiver.generateFluorophoresFromCSV(
                     this.file,
                     this.camera,
+                    this.laser,
                     this.psfBuilder,
                     this.fluorDynamics,
                     this.rescale);