+++ src/main/java/ch/epfl/leb/sass/models/fluorophores/internal/commands/GenerateFluorophoresGrid2D.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Laboratory of Experimental Biophysics
+ * Copyright (C) 2017-2018 Laboratory of Experimental Biophysics
  * Ecole Polytechnique Federale de Lausanne
  * 
  * This program is free software: you can redistribute it and/or modify
@@ -17,10 +17,12 @@
  */
 package ch.epfl.leb.sass.models.fluorophores.internal.commands;
 
+import ch.epfl.leb.sass.models.psfs.PSFBuilder;
+import ch.epfl.leb.sass.models.components.Laser;
+import ch.epfl.leb.sass.models.components.Camera;
 import ch.epfl.leb.sass.models.fluorophores.Fluorophore;
 import ch.epfl.leb.sass.models.photophysics.FluorophoreDynamics;
-import ch.epfl.leb.sass.models.components.Camera;
-import ch.epfl.leb.sass.models.psfs.PSFBuilder;
+
 import java.util.List;
 
 /**
@@ -45,6 +47,11 @@ public final class GenerateFluorophoresGrid2D implements FluorophoreCommand {
     private final FluorophoreDynamics fluorDynamics;
     
     /**
+     * The laser that illuminates the fluorophores.
+     */
+    private final Laser laser;
+    
+    /**
      * A builder for creating PSFs.
      */
     private final PSFBuilder psfBuilder;
@@ -55,6 +62,7 @@ public final class GenerateFluorophoresGrid2D implements FluorophoreCommand {
     public static class Builder implements FluorophoreCommandBuilder {
         private int spacing;
         private Camera camera;
+        private Laser laser;
         private FluorophoreDynamics fluorDynamics;
         private PSFBuilder psfBuilder;
         
@@ -70,6 +78,11 @@ public final class GenerateFluorophoresGrid2D implements FluorophoreCommand {
         @Override
         public Builder fluorDynamics(FluorophoreDynamics fluorDynamics) {
             this.fluorDynamics = fluorDynamics;
+            return this;
+        }
+        @Override
+        public Builder laser(Laser laser) {
+            this.laser = laser;
             return this;
         }
         @Override
@@ -91,6 +104,7 @@ public final class GenerateFluorophoresGrid2D implements FluorophoreCommand {
     private GenerateFluorophoresGrid2D(Builder builder) {
         this.camera = builder.camera;
         this.fluorDynamics = builder.fluorDynamics;
+        this.laser = builder.laser;
         this.spacing = builder.spacing;
         this.psfBuilder = builder.psfBuilder;
     }
@@ -105,6 +119,7 @@ public final class GenerateFluorophoresGrid2D implements FluorophoreCommand {
         return FluorophoreReceiver.generateFluorophoresGrid2D(
                 this.spacing, 
                 this.camera,
+                this.laser,
                 this.psfBuilder,
                 this.fluorDynamics);        
     }