+++ src/test/java/ch/epfl/leb/sass/server/RPCServerIT.java
@@ -134,6 +134,7 @@ public class RPCServerIT {
         laserBuilder.currentPower(0.0);
         laserBuilder.minPower(0.0);
         laserBuilder.maxPower(500.0);
+        laserBuilder.wavelength(0.642);
 
         // DefaultStage
         DefaultStage.Builder stageBuilder = new DefaultStage.Builder();